Create README.md
This is an open source quiz app, and its still in early development. Its intended to be used for fun and education about bad effects of food additives. The code was made this way so anyone could make a list of questions and  learning cards by modifying xml files without any knowledge of programming. I primary need help with localizing questions for different languages  and making new questions but  feel free to contribute in any way.